# Reuse hash in instance.Push to reduce memory allocations