@joe-elliott 2020-12-02
One other concern.  This change splits the hash generation method used in `FindTraceByID()` from the one used in `getOrCreateTrace()`.  I think we should change them to both use the same one.