@mdisibio 2020-12-03
> One other concern. This change splits the hash generation method used in `FindTraceByID()` from the one used in `getOrCreateTrace()`. I think we should change them to both use the same one.

Updated and added the benchmark results above, showing the elimination of 1 alloc/op.